separate asserts per php version
